trac#11494: fixes for workaround in FormGUI for SunJRE 1.6-Bug and restricted workaround execution
FormGUI contains a workaround for a Sun-JRE bug. This bug is suffering Sun-JREs < 1.7 and is no longer required for newer JRE-Versions (including OpenJDK). For this reason, I added the workaround to our Workarounds-Class and restricted it's calling for the suffered JRE's only.

Within the workaround there is a check if the FormGUI-window was successfully maximized. On LiMux BC 5.0 this check was no longer working correct. This caused the little state machine in the workaround to not switch automatically to the next correct state. It seems that for some strange reason (and only on some physical LiMux PCs) the LiMux BC 5.0 doesn't report Frame.MAXIMIZED_BOTH, but only Frame.MAXIMIZED_HORIZ. I made this check less restrictive to also recognize Frame.MAXIMIZED_HORIZ or Frame.MAXIMIZED_VERT actions. This fix was required to use the workaround in the less important case of running WollMux on BC 5.0 with Sun-JRE 1.6 (instead of the default openjdk 1.7)